python/requirements.txt
py-clob-client
catboost
xgboost
lightgbm
optuna
numpy
pandas
pyarrow
requests
websockets
python-socks[asyncio]
python-dotenv
grpcio
grpcio-tools
scikit-learn
matplotlib
shap
ta-lib
jupyter
